refac: move physic body state related to one file.
