--- src/core/PhysicBodyConfig.ts
@@ -1,10 +1,0 @@
-import {PhysicBodyType} from "../core-physic/PhysicBody.ts";
-import PhysicBodyProps from "./PhysicBodyProps.ts";
-import RendererProps from "../body-renderer/RendererProps.ts";
-
-export default interface PhysicBodyConfig {
-    type: PhysicBodyType;
-    dependencies?: {[key: string]: string};
-    props: PhysicBodyProps;
-    renderer?: RendererProps;
-}
