--- src/core/PhysicBodyProps.ts
@@ -1,3 +1,0 @@
-export default interface PhysicBodyProps {
-
-}
